fix(hydroflow_plus): handle send_bincode with local structs
Fixes #1144
