+++ hydroflow_plus_test/Cargo.toml
@@ -10,6 +10,7 @@ tokio = { version = "1.16", features = [ "full" ] }
 stageleft = { path = "../stageleft", version = "^0.2.1" }
 hydroflow_plus_cli_integration = { path = "../hydro_deploy/hydroflow_plus_cli_integration", version = "^0.6.1" }
 rand = "0.8.5"
+serde = { version = "1", features = [ "derive" ] }
 
 hydroflow_plus_test_macro = { path = "../hydroflow_plus_test_macro" }
